requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
pydantic
python-dotenv
